Renamed BaekhoBot to HypatIA and updated related references throughout the codebase. Adjusted agent services to reflect the new HypatIA branding and functionality.

+++ app/models/promocion/PromocionModel.py
@@ -1,11 +1,10 @@
 from pydantic import BaseModel, Field
 from typing import Optional
 from datetime import date
-from decimal import Decimal
 
 class PromocionBase(BaseModel):
     descripcion: str
-    descuentoPorcentaje: Decimal = Field(..., ge=0, le=100)
+    descuentoPorcentaje: int = Field(..., ge=0, le=100)
     fechaInicio: date
     fechaFin: date
 
@@ -14,7 +13,7 @@ class PromocionCreate(PromocionBase):
 
 class PromocionUpdate(BaseModel):
     descripcion: Optional[str] = None
-    descuentoPorcentaje: Optional[Decimal] = Field(None, ge=0, le=100)
+    descuentoPorcentaje: Optional[int] = Field(None, ge=0, le=100)
     fechaInicio: Optional[date] = None
     fechaFin: Optional[date] = None
